+++ idea/testData/refactoring/rename/automaticRenamerKotlinTestClass/after/test/JFooBarNewBazTestCase.java
@@ -1,0 +1,7 @@
+package test;
+
+import junit.framework.TestCase;
+
+public class JFooBarNewBazTestCase extends TestCase {
+
+}