+++ idea/testData/refactoring/rename/automaticRenamerKotlinTestClass/before/test/test.kt
@@ -1,0 +1,16 @@
+package test
+
+import junit.framework.TestCase
+import org.junit.Test
+
+class /*rename*/Bar {
+
+}
+
+class FooBarBazTestCase : TestCase() {
+
+}
+
+class BazBarFooTest {
+    @Test fun testFoo() {}
+}